Create compiler folder
