+++ pyteal/__init__.py
@@ -2,12 +2,15 @@ from .ast import *
 from .ast import __all__ as ast_all
 from .ir import *
 from .ir import __all__ as ir_all
-from .compiler import CompileOptions, compileTeal
+from .compiler import MAX_TEAL_VERSION, MIN_TEAL_VERSION, DEFAULT_TEAL_VERSION, CompileOptions, compileTeal
 from .types import TealType
 from .errors import TealInternalError, TealTypeError, TealInputError, TealCompileError
 from .config import MAX_GROUP_SIZE
 
 __all__ = ast_all + ir_all + [
+    "MAX_TEAL_VERSION",
+    "MIN_TEAL_VERSION",
+    "DEFAULT_TEAL_VERSION",
     "CompileOptions",
     "compileTeal",
     "TealType",
