+++ pyteal/compiler/__init__.py
@@ -1,0 +1,9 @@
+from .compiler import MAX_TEAL_VERSION, MIN_TEAL_VERSION, DEFAULT_TEAL_VERSION, CompileOptions, compileTeal
+
+__all__ = [
+    "MAX_TEAL_VERSION",
+    "MIN_TEAL_VERSION",
+    "DEFAULT_TEAL_VERSION",
+    "CompileOptions",
+    "compileTeal",
+]
