+++ pyteal/util.py
@@ -1,3 +1,4 @@
+from .errors import TealInternalError
 
 def escapeStr(s: str) -> str:
     """Escape a UTF-8 string for use in TEAL assembly.
@@ -29,3 +30,20 @@ def escapeStr(s: str) -> str:
     
     # Surround string in double quotes
     return "\"" + s + "\""
+
+def correctBase32Padding(s: str) -> str:
+    content = s.split('=')[0]
+    trailing = len(content) % 8
+
+    if trailing == 2:
+        content += '='*6
+    elif trailing == 4:
+        content += '='*4
+    elif trailing == 5:
+        content += '='*3
+    elif trailing == 7:
+        content += '='
+    elif trailing != 0:
+        raise TealInternalError("Invalid base32 content")
+
+    return content
